requirements.txt
aioconsole==0.3.3
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
backports.entry-points-selectable==1.1.1
bidict==0.21.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
distlib==0.3.3
filelock==3.4.0
Flask==2.0.2
Flask-SocketIO==5.1.1
frozenlist==1.2.0
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
numpy==1.21.4
platformdirs==2.4.0
python-engineio==4.3.0
python-socketio==5.5.0
requests==2.26.0
six==1.16.0
source==1.2.0
typing_extensions==4.0.0
urllib3==1.26.7
virtualenv==20.10.0
websocket-client==1.2.1
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0
